+++ geotrek/trekking/translation.py
@@ -11,7 +11,7 @@ class TrekTO(TranslationOptions):
     fields = ('name', 'departure', 'arrival', 'description_teaser',
               'description', 'ambiance', 'access', 'disabled_infrastructure', 'advice',
               'advised_parking', 'public_transport') + (
-              ('published',) if settings.TREK_PUBLISHED_BY_LANG else tuple()
+              ('published',) if settings.PUBLISHED_BY_LANG else tuple()
               )
     fallback_undefined = {
         'published': None
@@ -19,7 +19,9 @@ class TrekTO(TranslationOptions):
 
 
 class POITO(TranslationOptions):
-    fields = ('name', 'description', )
+    fields = ('name', 'description') + (
+              ('published',) if settings.PUBLISHED_BY_LANG else tuple()
+             )
 
 
 class POITypeTO(TranslationOptions):